# openapi.yaml
components:
  schemas: {}
info:
  title: ""
  version: 1.0.0
openapi: 3.0.0
paths:
  /users:
    get:
      operationId: UserController.getAll
      responses:
        "200":
          content:
            text/html; charset=utf-8: {}
          description: Successful response
      summary: Get all
      tags:
        - User
    post:
      operationId: UserController.post
      requestBody:
        content:
          application/json:
            schema: {}
        description: ""
        required: false
      responses:
        "200":
          content:
            text/html; charset=utf-8: {}
          description: Successful response
      summary: Post
      tags:
        - User
  "/users/{id}":
    get:
      operationId: UserController.getOne
      parameters:
        - in: path
          name: id
          required: true
          schema:
            pattern: "[^\\/#\\?]+?"
            type: number
        - in: query
          name: type
          required: true
          schema:
            $ref: "#/components/schemas/Number"
      responses:
        "200":
          content:
            text/html; charset=utf-8: {}
          description: Successful response
        "400":
          description: Bad request
      summary: Get one
      tags:
        - User
      description: Fucking fuck
    put:
      operationId: UserController.put
      parameters:
        - in: path
          name: id
          required: true
          schema:
            pattern: "[^\\/#\\?]+?"
            type: number
      requestBody:
        content:
          application/json:
            schema:
              properties:
                user:
                  $ref: "#/components/schemas/UserBody"
              required:
                - user
              type: object
      responses:
        "200":
          content:
            application/json: {}
          description: Successful response
      summary: Put
      tags:
        - User
    delete:
      operationId: UserController.remove
      parameters:
        - in: path
          name: id
          required: true
          schema:
            pattern: "[^\\/#\\?]+?"
            type: number
      responses:
        "200":
          content:
            text/html; charset=utf-8: {}
          description: Successful response
      summary: Remove
      tags:
        - User